refactor: implement comprehensive gradient system with mathematical accuracy
- Add VjpCompute trait with multiple method variants for different operation types
- Fix critical mathematical errors in gradient computations:
  * Sigmoid: correct 1-sigmoid formula (was computing sigmoid-1)
  * GELU: implement full mathematical derivative replacing simple approximation
- Implement dimension-aware reduce operation gradients:
  * Add compute_vjp_with_dims method for proper dimensional handling
  * Fix Mean/Var/Std to use actual reduced dimensions instead of all dimensions
- Enhance transpose gradient computation with automatic dimension inference
- Add comprehensive error handling for division by zero in scalar operations
- Change transpose methods to use i32 instead of isize for consistency
- Remove unused TensorOp enum and clean up Op structure
- Add utility functions for gradient computations

+++ crates/hodu_core/src/tensor/ops.rs
@@ -934,7 +934,7 @@ impl Tensor {
         self.broadcast(&target_shape)
     }
 
-    pub fn transpose(&self, dim1: isize, dim2: isize) -> HoduResult<Self> {
+    pub fn transpose(&self, dim1: i32, dim2: i32) -> HoduResult<Self> {
         let layout = self.get_layout();
         let new_layout = layout.transpose(dim1, dim2)?;
         let requires_grad = self.is_requires_grad();
